Update splitter.py
How did the ability to drag and drop files from any location into the list box.
Also added the ability to specify the number of images wide and tall the character sheet is.
The script will automatically detect the ratio needed for the Split Image as well. To avoid any stretching or deforming of the smaller images.